velog/requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.51.0
django-extensions==3.2.1
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
exceptiongroup==1.0.4
filelock==3.8.2
gitmoji==0.1.0
identify==2.5.11
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
isort==5.11.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
model-mommy==2.0.0
mypy-extensions==0.4.3
nodeenv==1.7.0
oauthlib==3.2.2
packaging==22.0
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.1
pluggy==1.0.0
pre-commit==2.21.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
toposort==1.7
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.1
